refactor: improve web tools extraction and WebFetch presentation
- split web tools into a dedicated module and re-export WebSearch/WebFetch
- add readable HTML extraction with markdown-first output and plain-text fallback
- return extracted page title and related metadata from WebFetch responses
- document extractor tradeoffs and keep rs-trafilatura disabled for now
- polish WebFetch card with title-first header, expanded metadata, and copy action
- update focused tests and locale strings for the new WebFetch UX
