+++ src/crates/assembly/core/src/agentic/tools/implementations/web/search.rs
@@ -1,0 +1,363 @@
+use crate::agentic::tools::framework::{Tool, ToolExposure, ToolResult, ToolUseContext};
+use crate::util::errors::{BitFunError, BitFunResult};
+use async_trait::async_trait;
+use log::{error, info};
+use serde::Deserialize;
+use serde_json::{json, Value};
+use std::time::Duration;
+
+const EXA_URL: &str = "https://mcp.exa.ai/mcp";
+const EXA_RESULTS: u64 = 5;
+const EXA_CONTEXT: u64 = 8_000;
+
+#[derive(Debug, Deserialize)]
+struct ExaRes {
+    result: Option<ExaData>,
+}
+
+#[derive(Debug, Deserialize)]
+struct ExaData {
+    content: Vec<ExaContent>,
+}
+
+#[derive(Debug, Deserialize)]
+struct ExaContent {
+    #[serde(rename = "type")]
+    kind: String,
+    text: Option<String>,
+}
+
+pub struct WebSearchTool;
+
+impl Default for WebSearchTool {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl WebSearchTool {
+    pub fn new() -> Self {
+        Self
+    }
+
+    async fn search(
+        &self,
+        query: &str,
+        num: u64,
+        kind: &str,
+        crawl: &str,
+        ctx: u64,
+    ) -> BitFunResult<String> {
+        let cli = reqwest::Client::builder()
+            .timeout(Duration::from_secs(25))
+            .build()
+            .map_err(|err| BitFunError::tool(format!("Failed to create HTTP client: {}", err)))?;
+
+        let body = json!({
+            "jsonrpc": "2.0",
+            "id": 1,
+            "method": "tools/call",
+            "params": {
+                "name": "web_search_exa",
+                "arguments": {
+                    "query": query,
+                    "type": kind,
+                    "numResults": num,
+                    "livecrawl": crawl,
+                    "contextMaxCharacters": ctx,
+                }
+            }
+        });
+
+        let res = cli
+            .post(EXA_URL)
+            .header("accept", "application/json, text/event-stream")
+            .header("content-type", "application/json")
+            .json(&body)
+            .send()
+            .await
+            .map_err(|err| BitFunError::tool(format!("Failed to send request: {}", err)))?;
+
+        let status = res.status();
+        if !status.is_success() {
+            let err = res
+                .text()
+                .await
+                .unwrap_or_else(|_| String::from("Unknown error"));
+            error!("WebSearch Exa error: status={}, error={}", status, err);
+            return Err(BitFunError::tool(format!(
+                "Web search error {}: {}",
+                status, err
+            )));
+        }
+
+        let text = res
+            .text()
+            .await
+            .map_err(|err| BitFunError::tool(format!("Failed to read response: {}", err)))?;
+
+        self.parse_sse(&text)
+    }
+
+    fn parse_sse(&self, text: &str) -> BitFunResult<String> {
+        let out = text
+            .lines()
+            .filter_map(|line| line.strip_prefix("data: "))
+            .find_map(|line| {
+                serde_json::from_str::<ExaRes>(line)
+                    .ok()
+                    .and_then(|res| res.result)
+                    .map(|res| {
+                        res.content
+                            .into_iter()
+                            .filter(|item| item.kind == "text")
+                            .filter_map(|item| item.text)
+                            .collect::<Vec<_>>()
+                            .join("\n")
+                    })
+                    .filter(|item| !item.trim().is_empty())
+            });
+
+        out.ok_or_else(|| BitFunError::tool("Web search returned no content".to_string()))
+    }
+
+    pub(crate) fn results(&self, text: &str) -> Vec<Value> {
+        let mut out = Vec::new();
+        let mut cur: Option<(String, String, Vec<String>)> = None;
+        let mut body = false;
+
+        for line in text.lines() {
+            if let Some(next) = line.strip_prefix("Title: ") {
+                if let Some((title, url, text)) = cur.take() {
+                    out.push(self.item(title, url, text));
+                }
+                cur = Some((next.trim().to_string(), String::new(), Vec::new()));
+                body = false;
+                continue;
+            }
+
+            let Some(cur) = cur.as_mut() else {
+                continue;
+            };
+
+            if let Some(next) = line.strip_prefix("URL: ") {
+                cur.1 = next.trim().to_string();
+                continue;
+            }
+
+            if let Some(next) = line.strip_prefix("Text: ") {
+                if !next.trim().is_empty() {
+                    cur.2.push(next.trim().to_string());
+                }
+                body = true;
+                continue;
+            }
+
+            if body {
+                cur.2.push(line.to_string());
+            }
+        }
+
+        if let Some((title, url, text)) = cur.take() {
+            out.push(self.item(title, url, text));
+        }
+
+        if out.is_empty() && !text.trim().is_empty() {
+            return vec![json!({
+                "title": "Web search result",
+                "url": "",
+                "snippet": self.snippet(text)
+            })];
+        }
+
+        out
+    }
+
+    fn item(&self, title: String, url: String, text: Vec<String>) -> Value {
+        json!({
+            "title": title,
+            "url": url,
+            "snippet": self.snippet(&text.join("\n"))
+        })
+    }
+
+    fn snippet(&self, text: &str) -> String {
+        let text = text
+            .lines()
+            .map(str::trim)
+            .filter(|line| !line.is_empty())
+            .filter(|line| !line.starts_with('#'))
+            .collect::<Vec<_>>()
+            .join(" ");
+
+        if text.chars().count() <= 320 {
+            return text;
+        }
+
+        let mut out = String::new();
+        for ch in text.chars().take(317) {
+            out.push(ch);
+        }
+        out.push_str("...");
+        out
+    }
+}
+
+#[async_trait]
+impl Tool for WebSearchTool {
+    fn name(&self) -> &str {
+        "WebSearch"
+    }
+
+    async fn description(&self) -> BitFunResult<String> {
+        Ok(
+            r#"- Allows BitFun to search the web and use the results to inform responses
+- Provides up-to-date information for current events and recent data
+- Returns search result information formatted as search result blocks
+- Use this tool for accessing information beyond BitFun's knowledge cutoff
+
+Usage notes:
+- Use when you need current information not in training data
+- Effective for recent news, current events, product updates, or real-time data
+- Search queries should be specific and well-targeted for best results
+- Results include title, URL, snippet and source information
+
+Advanced features:
+- Choose search depth: auto, fast, or deep
+- Control result count and context size for LLM-friendly output
+- Optionally prefer live crawling for fresher pages
+- Return up to 10 results per query"#
+                .to_string(),
+        )
+    }
+
+    fn short_description(&self) -> String {
+        "Search the web for up-to-date information and sources.".to_string()
+    }
+
+    fn default_exposure(&self) -> ToolExposure {
+        ToolExposure::Collapsed
+    }
+
+    fn input_schema(&self) -> Value {
+        json!({
+            "type": "object",
+            "properties": {
+                "query": {
+                    "type": "string",
+                    "description": "The search query (recommended max 70 characters)"
+                },
+                "num_results": {
+                    "type": "number",
+                    "description": "Number of search results to return (1-10, default: 5)",
+                    "default": EXA_RESULTS,
+                    "minimum": 1,
+                    "maximum": 10
+                },
+                "type": {
+                    "type": "string",
+                    "enum": ["auto", "fast", "deep"],
+                    "description": "Search depth. Use 'auto' for balanced results, 'fast' for lower latency, or 'deep' for broader context.",
+                    "default": "auto"
+                },
+                "livecrawl": {
+                    "type": "string",
+                    "enum": ["fallback", "preferred"],
+                    "description": "Live crawl mode. Use 'preferred' to favor fresh crawling, or 'fallback' to use cached data when possible.",
+                    "default": "fallback"
+                },
+                "context_max_characters": {
+                    "type": "number",
+                    "description": "Maximum characters of search context to request (default: 8000)",
+                    "default": EXA_CONTEXT,
+                    "minimum": 1000,
+                    "maximum": 20000
+                }
+            },
+            "required": ["query"]
+        })
+    }
+
+    fn is_readonly(&self) -> bool {
+        true
+    }
+
+    fn is_concurrency_safe(&self, _input: Option<&Value>) -> bool {
+        true
+    }
+
+    fn needs_permissions(&self, _input: Option<&Value>) -> bool {
+        false
+    }
+
+    async fn call_impl(
+        &self,
+        input: &Value,
+        _context: &ToolUseContext,
+    ) -> BitFunResult<Vec<ToolResult>> {
+        let query = input
+            .get("query")
+            .and_then(|v| v.as_str())
+            .ok_or_else(|| BitFunError::tool("query is required".to_string()))?;
+
+        let num_results = input
+            .get("num_results")
+            .and_then(|v| v.as_u64())
+            .unwrap_or(EXA_RESULTS)
+            .clamp(1, 10);
+
+        let kind = input.get("type").and_then(|v| v.as_str()).unwrap_or("auto");
+
+        let crawl = input
+            .get("livecrawl")
+            .and_then(|v| v.as_str())
+            .unwrap_or("fallback");
+
+        let ctx = input
+            .get("context_max_characters")
+            .and_then(|v| v.as_u64())
+            .unwrap_or(EXA_CONTEXT)
+            .clamp(1_000, 20_000);
+
+        info!(
+            "WebSearch Exa call: query='{}', num_results={}, type={}, livecrawl={}, context_max_characters={}",
+            query, num_results, kind, crawl, ctx
+        );
+
+        let raw = self.search(query, num_results, kind, crawl, ctx).await?;
+        let results = self.results(&raw);
+
+        let formatted_results = results
+            .iter()
+            .enumerate()
+            .map(|(i, r)| {
+                format!(
+                    "{}. {}\n   URL: {}\n   Snippet: {}\n",
+                    i + 1,
+                    r["title"].as_str().unwrap_or("Untitled"),
+                    r["url"].as_str().unwrap_or(""),
+                    r["snippet"].as_str().unwrap_or("")
+                )
+            })
+            .collect::<Vec<_>>()
+            .join("\n");
+
+        let result = ToolResult::Result {
+            data: json!({
+                "query": query,
+                "results": results,
+                "result_count": results.len(),
+                "provider": "exa_mcp"
+            }),
+            result_for_assistant: Some(format!(
+                "Search query: '{}'\nFound {} results:\n\n{}",
+                query,
+                results.len(),
+                formatted_results
+            )),
+            image_attachments: None,
+        };
+
+        Ok(vec![result])
+    }
+}
