+++ src/web-ui/src/locales/en-US/flow-chat.json
@@ -1413,10 +1413,14 @@
     },
     "webFetch": {
       "parsingUrl": "Parsing URL...",
+      "readLabel": "Read Webpage:",
       "readTitle": "Read Webpage: \"{{url}}\"",
       "reading": "Reading \"{{url}}\"...",
       "preparingRead": "Preparing to read \"{{url}}\"",
       "clickToOpenLink": "Click to open link",
+      "copyResult": "Copy result",
+      "resultCopied": "Result copied",
+      "copyResultFailed": "Failed to copy result",
       "contentAvailable": "content available",
       "contentLength": "{{count}} chars",
       "fetchFailed": "Failed to fetch content",
