+++ src/web-ui/src/locales/zh-CN/flow-chat.json
@@ -1413,10 +1413,14 @@
     },
     "webFetch": {
       "parsingUrl": "解析链接中...",
+      "readLabel": "读取网页:",
       "readTitle": "读取网页: \"{{url}}\"",
       "reading": "正在读取 \"{{url}}\"...",
       "preparingRead": "准备读取 \"{{url}}\"",
       "clickToOpenLink": "点击打开链接",
+      "copyResult": "复制结果",
+      "resultCopied": "结果已复制",
+      "copyResultFailed": "复制结果失败",
       "contentAvailable": "有内容可预览",
       "contentLength": "{{count}} 个字符",
       "fetchFailed": "抓取内容失败",
