+++ src/web-ui/src/locales/zh-TW/flow-chat.json
@@ -1413,10 +1413,14 @@
     },
     "webFetch": {
       "parsingUrl": "解析連結中...",
+      "readLabel": "讀取網頁:",
       "readTitle": "讀取網頁: \"{{url}}\"",
       "reading": "正在讀取 \"{{url}}\"...",
       "preparingRead": "準備讀取 \"{{url}}\"",
       "clickToOpenLink": "點擊開啟鏈接",
+      "copyResult": "複製結果",
+      "resultCopied": "結果已複製",
+      "copyResultFailed": "複製結果失敗",
       "contentAvailable": "有內容可預覽",
       "contentLength": "{{count}} 個字符",
       "fetchFailed": "抓取內容失敗",
